test: split large test files + fix build tags
