+++ tests/integration/gitserver_test.go
@@ -1,3 +1,5 @@
+//go:build integration
+
 package integration
 
 import (
