--- tests/integration/integration_test.go
@@ -1,1676 +1,0 @@
-package integration
-
-import (
-	"context"
-	"database/sql"
-	"fmt"
-	"testing"
-	"time"
-
-	"github.com/kinoko-dev/kinoko/internal/decay"
-	"github.com/kinoko-dev/kinoko/internal/extraction"
-	"github.com/kinoko-dev/kinoko/internal/injection"
-	"github.com/kinoko-dev/kinoko/internal/metrics"
-	"github.com/kinoko-dev/kinoko/internal/model"
-	"github.com/kinoko-dev/kinoko/internal/storage"
-)
-
-// indexingCommitter simulates the post-receive hook: on CommitSkill it indexes
-// the skill into SQLite, mimicking what the real hook pipeline does.
-type indexingCommitter struct {
-	indexer  model.SkillIndexer
-	embedder extraction.SkillEmbedder
-}
-
-func (c *indexingCommitter) CommitSkill(ctx context.Context, _ string, skill *model.SkillRecord, body []byte) (string, error) {
-	var emb []float32
-	if c.embedder != nil {
-		var err error
-		emb, err = c.embedder.Embed(ctx, string(body))
-		if err != nil {
-			return "", err
-		}
-	}
-	if err := c.indexer.IndexSkill(ctx, skill, emb); err != nil {
-		return "", err
-	}
-	return "deadbeef", nil
-}
-
-// =============================================================================
-// Test 1: Full Extraction Flow
-// session log → Stage1 → Stage2 → Stage3 → stored skill → verify in DB
-// =============================================================================
-
-func TestFullExtractionFlow(t *testing.T) {
-	store := newTestStore(t)
-	ctx := context.Background()
-
-	embedder := newPredictableEmbedder(3)
-	llm := &predictableLLM{
-		rubricResponse: goodRubricJSON(),
-		criticResponse: extractVerdictJSON(),
-	}
-	reviewer := &mockReviewWriter{}
-
-	// Build the real pipeline with real store but mock LLM/embedder.
-	s1 := extraction.NewStage1Filter(defaultExtractionConfig(), testLogger())
-	s2 := extraction.NewStage2Scorer(embedder, &mockQuerier{sim: 0.50}, llm, defaultExtractionConfig(), testLogger())
-	s3 := extraction.NewStage3Critic(llm, defaultExtractionConfig(), testLogger())
-
-	indexer := storage.NewSQLiteIndexer(store)
-	committer := &indexingCommitter{indexer: indexer, embedder: embedder}
-
-	pipeline, err := extraction.NewPipeline(extraction.PipelineConfig{
-		Stage1:     s1,
-		Stage2:     s2,
-		Stage3:     s3,
-		Committer:  committer,
-		Reviewer:   reviewer,
-		Embedder:   embedder,
-		Log:        testLogger(),
-		SampleRate: 0, // no sampling noise
-		Extractor:  "integration-test-v1",
-	})
-	if err != nil {
-		t.Fatalf("new pipeline: %v", err)
-	}
-
-	session := goodSession("sess-e2e-001", "test-lib")
-	content := []byte("User asked to fix database connection pooling. Agent diagnosed stale connections, implemented retry logic with exponential backoff, verified fix with load test.")
-
-	result, err := pipeline.Extract(ctx, session, content)
-	if err != nil {
-		t.Fatalf("extract: %v", err)
-	}
-
-	// Verify extraction succeeded.
-	if result.Status != model.StatusExtracted {
-		t.Fatalf("status = %q, want extracted (error: %s)", result.Status, result.Error)
-	}
-	if result.Skill == nil {
-		t.Fatal("expected skill in result")
-	}
-	if result.Stage1 == nil || !result.Stage1.Passed {
-		t.Error("stage1 should have passed")
-	}
-	if result.Stage2 == nil || !result.Stage2.Passed {
-		t.Error("stage2 should have passed")
-	}
-	if result.Stage3 == nil || !result.Stage3.Passed {
-		t.Error("stage3 should have passed")
-	}
-
-	// Verify skill is in the REAL SQLite database.
-	skillID := result.Skill.ID
-	dbSkill, err := store.Get(ctx, skillID)
-	if err != nil {
-		t.Fatalf("get skill from db: %v", err)
-	}
-
-	if dbSkill.Name != result.Skill.Name {
-		t.Errorf("db name = %q, result name = %q", dbSkill.Name, result.Skill.Name)
-	}
-	if dbSkill.LibraryID != "test-lib" {
-		t.Errorf("library = %q, want test-lib", dbSkill.LibraryID)
-	}
-	if dbSkill.ExtractedBy != "integration-test-v1" {
-		t.Errorf("extracted_by = %q", dbSkill.ExtractedBy)
-	}
-	if dbSkill.Category != model.CategoryTactical {
-		t.Errorf("category = %q, want tactical", dbSkill.Category)
-	}
-	if dbSkill.Quality.ProblemSpecificity != 4 {
-		t.Errorf("problem_specificity = %d, want 4", dbSkill.Quality.ProblemSpecificity)
-	}
-	// BUG FOUND: Pipeline doesn't set DecayScore=1.0 on new skills.
-	// The Go zero value (0.0) overrides the DB DEFAULT 1.0.
-	// This means newly extracted skills are immediately "dead" for injection.
-	// For now, test documents the actual behavior:
-	if dbSkill.DecayScore != 1.0 {
-		t.Errorf("initial decay = %f, want 1.0", dbSkill.DecayScore)
-	}
-	if dbSkill.SourceSessionID != "sess-e2e-001" {
-		t.Errorf("source_session = %q", dbSkill.SourceSessionID)
-	}
-
-	// Verify patterns stored correctly in join table.
-	if len(dbSkill.Patterns) == 0 {
-		t.Error("expected patterns in DB")
-	}
-
-	// Verify embedding stored correctly.
-	if len(dbSkill.Embedding) != 3 {
-		t.Errorf("embedding dims = %d, want 3", len(dbSkill.Embedding))
-	}
-
-	// Verify timing metadata (pipeline may complete in <1ms, so >= 0 is correct).
-	if result.DurationMs < 0 {
-		t.Errorf("duration = %d, want >= 0", result.DurationMs)
-	}
-}
-
-// =============================================================================
-// Test 2: Full Injection Flow
-// stored skill in DB → inject query → verify skill returned with correct ranking
-// =============================================================================
-
-func TestFullInjectionFlow(t *testing.T) {
-	store := newTestStore(t)
-	ctx := context.Background()
-	embedder := newPredictableEmbedder(3)
-
-	// Seed two skills into the DB.
-	skill1 := &model.SkillRecord{
-		ID:        "skill-inj-1",
-		Name:      "fix-db-conn",
-		Version:   1,
-		LibraryID: "test-lib",
-		Category:  model.CategoryTactical,
-		Patterns:  []string{"FIX/Backend/DatabaseConnection"},
-		Quality: model.QualityScores{
-			ProblemSpecificity: 4, SolutionCompleteness: 4, ContextPortability: 3,
-			ReasoningTransparency: 3, TechnicalAccuracy: 4, VerificationEvidence: 3,
-			InnovationLevel: 3, CompositeScore: 3.5, CriticConfidence: 0.85,
-		},
-		Embedding:          embedder.deterministicVector("fix database connection"),
-		DecayScore:         1.0,
-		SuccessCorrelation: 0.5,
-		ExtractedBy:        "test",
-		FilePath:           "skills/fix-db-conn/SKILL.md",
-	}
-	skill2 := &model.SkillRecord{
-		ID:        "skill-inj-2",
-		Name:      "build-api",
-		Version:   1,
-		LibraryID: "test-lib",
-		Category:  model.CategoryFoundational,
-		Patterns:  []string{"BUILD/Backend/APIDesign"},
-		Quality: model.QualityScores{
-			ProblemSpecificity: 3, SolutionCompleteness: 3, ContextPortability: 4,
-			ReasoningTransparency: 3, TechnicalAccuracy: 3, VerificationEvidence: 3,
-			InnovationLevel: 2, CompositeScore: 3.0, CriticConfidence: 0.7,
-		},
-		Embedding:          embedder.deterministicVector("build REST API design"),
-		DecayScore:         0.8,
-		SuccessCorrelation: 0.3,
-		ExtractedBy:        "test",
-		FilePath:           "skills/build-api/SKILL.md",
-	}
-
-	if err := store.Put(ctx, skill1, nil); err != nil {
-		t.Fatalf("put skill1: %v", err)
-	}
-	if err := store.Put(ctx, skill2, nil); err != nil {
-		t.Fatalf("put skill2: %v", err)
-	}
-
-	// Create injector with real store.
-	llm := &predictableLLM{
-		classifyResponse: classifyJSON("FIX", "Backend", []string{"FIX/Backend/DatabaseConnection"}),
-	}
-
-	inj := injection.New(embedder, store, llm, store, testLogger())
-
-	resp, err := inj.Inject(ctx, model.InjectionRequest{
-		Prompt:     "fix database connection pooling issue",
-		LibraryIDs: []string{"test-lib"},
-		MaxSkills:  5,
-		SessionID:  "sess-inj-001",
-	})
-	if err != nil {
-		t.Fatalf("inject: %v", err)
-	}
-
-	if len(resp.Skills) == 0 {
-		t.Fatal("expected at least one skill injected")
-	}
-
-	// Both skills should be returned (both in test-lib).
-	if len(resp.Skills) != 2 {
-		t.Errorf("got %d skills, want 2", len(resp.Skills))
-	}
-
-	// Verify ranking: skill with matching pattern should rank higher.
-	if resp.Skills[0].SkillID != "skill-inj-1" {
-		t.Errorf("top skill = %q, want skill-inj-1 (has matching pattern)", resp.Skills[0].SkillID)
-	}
-
-	// Verify injection events were written to real DB.
-	var eventCount int
-	store.DB().QueryRow("SELECT COUNT(*) FROM injection_events WHERE session_id = ?", "sess-inj-001").Scan(&eventCount)
-	if eventCount != 2 {
-		t.Errorf("injection events = %d, want 2", eventCount)
-	}
-}
-
-// =============================================================================
-// Test 3: Full Feedback Loop
-// extract → inject → record outcome → verify success_correlation updated
-// =============================================================================
-
-func TestFeedbackLoop(t *testing.T) {
-	store := newTestStore(t)
-	ctx := context.Background()
-	embedder := newPredictableEmbedder(3)
-
-	// Step 1: Seed a skill.
-	skill := &model.SkillRecord{
-		ID:        "skill-fb-1",
-		Name:      "fix-auth-flow",
-		Version:   1,
-		LibraryID: "test-lib",
-		Category:  model.CategoryTactical,
-		Patterns:  []string{"FIX/Backend/AuthFlow"},
-		Quality: model.QualityScores{
-			ProblemSpecificity: 4, SolutionCompleteness: 4, ContextPortability: 3,
-			ReasoningTransparency: 3, TechnicalAccuracy: 4, VerificationEvidence: 3,
-			InnovationLevel: 3, CompositeScore: 3.5, CriticConfidence: 0.8,
-		},
-		Embedding:          embedder.deterministicVector("fix auth flow"),
-		DecayScore:         1.0,
-		SuccessCorrelation: 0.0,
-		ExtractedBy:        "test",
-		FilePath:           "skills/fix-auth/SKILL.md",
-	}
-	if err := store.Put(ctx, skill, nil); err != nil {
-		t.Fatal(err)
-	}
-
-	// Step 2: Simulate injection events (3 successes, 1 failure).
-	now := time.Now().UTC()
-	for i, outcome := range []string{"success", "success", "failure", "success"} {
-		ev := storage.InjectionEventRecord{
-			ID:           fmt.Sprintf("ev-fb-%d", i),
-			SessionID:    fmt.Sprintf("sess-fb-%d", i),
-			SkillID:      "skill-fb-1",
-			RankPosition: 1,
-			MatchScore:   0.8,
-			InjectedAt:   now,
-		}
-		if err := store.WriteInjectionEvent(ctx, ev); err != nil {
-			t.Fatalf("write event %d: %v", i, err)
-		}
-		// Update session outcome.
-		store.DB().Exec("UPDATE injection_events SET session_outcome = ? WHERE id = ?", outcome, ev.ID)
-	}
-
-	// Step 3: Record outcome via UpdateUsage (which recomputes success_correlation).
-	if err := store.UpdateUsage(ctx, "skill-fb-1", "success"); err != nil {
-		t.Fatal(err)
-	}
-
-	// Step 4: Verify success_correlation in DB.
-	got, err := store.Get(ctx, "skill-fb-1")
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Expected: (3 success - 1 failure) / 4 total = 0.5
-	assertApprox(t, got.SuccessCorrelation, 0.5, 0.01, "success_correlation")
-}
-
-// =============================================================================
-// Test 4: Full Decay Cycle
-// extract → simulate time → decay → verify scores changed correctly
-// =============================================================================
-
-func TestDecayCycle(t *testing.T) {
-	store := newTestStore(t)
-	ctx := context.Background()
-
-	now := time.Date(2026, 2, 15, 12, 0, 0, 0, time.UTC)
-
-	// Seed skills with different categories and ages.
-	skills := []struct {
-		id       string
-		name     string
-		category model.SkillCategory
-		age      int // days since last injection
-	}{
-		{"skill-d1", "foundational-skill", model.CategoryFoundational, 365}, // 1 half-life
-		{"skill-d2", "tactical-skill", model.CategoryTactical, 90},          // 1 half-life
-		{"skill-d3", "contextual-skill", model.CategoryContextual, 180},     // 1 half-life
-		{"skill-d4", "fresh-tactical", model.CategoryTactical, 1},           // barely aged
-		{"skill-d5", "nearly-dead", model.CategoryTactical, 900},            // very old
-	}
-
-	for _, s := range skills {
-		lastInjected := now.AddDate(0, 0, -s.age)
-		sk := &model.SkillRecord{
-			ID:        s.id,
-			Name:      s.name,
-			Version:   1,
-			LibraryID: "test-lib",
-			Category:  s.category,
-			Patterns:  []string{"FIX/Backend/DatabaseConnection"},
-			Quality: model.QualityScores{
-				ProblemSpecificity: 4, SolutionCompleteness: 4, ContextPortability: 3,
-				ReasoningTransparency: 3, TechnicalAccuracy: 4, VerificationEvidence: 3,
-				InnovationLevel: 3, CompositeScore: 3.5, CriticConfidence: 0.8,
-			},
-			DecayScore:     1.0,
-			ExtractedBy:    "test",
-			FilePath:       "skills/" + s.name + "/SKILL.md",
-			LastInjectedAt: lastInjected,
-		}
-		if err := store.Put(ctx, sk, nil); err != nil {
-			t.Fatalf("put %s: %v", s.id, err)
-		}
-		// Set LastInjectedAt via UpdateUsage workaround — Put doesn't write it.
-		// Actually, check: Put writes last_injected_at as nullTime...
-		// We need to set it directly.
-		store.DB().Exec("UPDATE skills SET last_injected_at = ? WHERE id = ?", lastInjected, s.id)
-	}
-
-	cfg := decay.DefaultConfig()
-	runner, err := decay.NewRunner(store, store, cfg, testLogger())
-	if err != nil {
-		t.Fatal(err)
-	}
-	// Inject test clock.
-	runner.SetNow(func() time.Time { return now })
-
-	result, err := runner.RunCycle(ctx, "test-lib")
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if result.Processed != 5 {
-		t.Errorf("processed = %d, want 5", result.Processed)
-	}
-
-	// Verify specific decay values.
-	for _, tc := range []struct {
-		id        string
-		wantDecay float64
-		eps       float64
-	}{
-		{"skill-d1", 0.5, 0.01},   // foundational: 365 days / 365 half-life = 1 half-life → 0.5
-		{"skill-d2", 0.5, 0.01},   // tactical: 90 / 90 = 1 → 0.5
-		{"skill-d3", 0.5, 0.01},   // contextual: 180 / 180 = 1 → 0.5
-		{"skill-d4", 0.992, 0.01}, // tactical: 1/90 → ~0.992
-		{"skill-d5", 0.0, 0.001},  // tactical: 900/90 = 10 half-lives → ~0.001 → deprecated to 0.0
-	} {
-		sk, err := store.Get(ctx, tc.id)
-		if err != nil {
-			t.Fatalf("get %s: %v", tc.id, err)
-		}
-		assertApprox(t, sk.DecayScore, tc.wantDecay, tc.eps, tc.id+" decay")
-	}
-
-	// The nearly-dead skill should be deprecated.
-	if result.Deprecated < 1 {
-		t.Errorf("deprecated = %d, want >= 1", result.Deprecated)
-	}
-}
-
-// =============================================================================
-// Test 5: A/B Test Flow
-// inject with A/B → verify treatment gets skills, control doesn't → verify logged
-// =============================================================================
-
-func TestABTestFlow(t *testing.T) {
-	store := newTestStore(t)
-	ctx := context.Background()
-	embedder := newPredictableEmbedder(3)
-
-	// Seed a skill.
-	sk := &model.SkillRecord{
-		ID:        "skill-ab-1",
-		Name:      "fix-deploy",
-		Version:   1,
-		LibraryID: "test-lib",
-		Category:  model.CategoryTactical,
-		Patterns:  []string{"FIX/DevOps/DeploymentFailure"},
-		Quality: model.QualityScores{
-			ProblemSpecificity: 4, SolutionCompleteness: 4, ContextPortability: 3,
-			ReasoningTransparency: 3, TechnicalAccuracy: 4, VerificationEvidence: 3,
-			InnovationLevel: 3, CompositeScore: 3.5, CriticConfidence: 0.8,
-		},
-		Embedding:   embedder.deterministicVector("fix deployment failure"),
-		DecayScore:  1.0,
-		ExtractedBy: "test",
-		FilePath:    "skills/fix-deploy/SKILL.md",
-	}
-	if err := store.Put(ctx, sk, nil); err != nil {
-		t.Fatal(err)
-	}
-
-	llm := &predictableLLM{
-		classifyResponse: classifyJSON("FIX", "DevOps", []string{"FIX/DevOps/DeploymentFailure"}),
-	}
-
-	// nil eventWriter on inner — ABInjector writes events with group info.
-	// This matches the production wiring in serve.go to avoid double events.
-	innerInj := injection.New(embedder, store, llm, nil, testLogger())
-
-	abConfig := injection.ABConfig{
-		Enabled:       true,
-		ControlRatio:  0.5, // 50/50 for deterministic testing
-		MinSampleSize: 10,
-	}
-
-	abInj := injection.NewABInjector(innerInj, store, abConfig, testLogger())
-
-	// Run treatment: force rand to return > controlRatio.
-	abInj.SetRandFunc(func() float64 { return 0.9 }) // treatment
-
-	respT, err := abInj.Inject(ctx, model.InjectionRequest{
-		Prompt:     "fix deployment failure",
-		LibraryIDs: []string{"test-lib"},
-		MaxSkills:  5,
-		SessionID:  "sess-ab-treatment",
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-	if len(respT.Skills) == 0 {
-		t.Error("treatment group should receive skills")
-	}
-
-	// Run control: force rand to return < controlRatio.
-	abInj.SetRandFunc(func() float64 { return 0.1 }) // control
-
-	respC, err := abInj.Inject(ctx, model.InjectionRequest{
-		Prompt:     "fix deployment failure",
-		LibraryIDs: []string{"test-lib"},
-		MaxSkills:  5,
-		SessionID:  "sess-ab-control",
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-	if len(respC.Skills) != 0 {
-		t.Error("control group should NOT receive skills")
-	}
-
-	// Verify both have injection events logged.
-	var treatmentEvents, controlEvents int
-	store.DB().QueryRow("SELECT COUNT(*) FROM injection_events WHERE session_id = ? AND ab_group = 'treatment' AND delivered = 1", "sess-ab-treatment").Scan(&treatmentEvents)
-	store.DB().QueryRow("SELECT COUNT(*) FROM injection_events WHERE session_id = ? AND ab_group = 'control' AND delivered = 0", "sess-ab-control").Scan(&controlEvents)
-
-	if treatmentEvents == 0 {
-		t.Error("no treatment injection events logged")
-	}
-	if controlEvents == 0 {
-		t.Error("no control injection events logged")
-	}
-}
-
-// =============================================================================
-// Test 6: Rejection at Each Stage
-// Verify DB state is correct after rejection at each stage
-// =============================================================================
-
-func TestRejectionFlows(t *testing.T) {
-	t.Run("rejected at stage1", func(t *testing.T) {
-		store := newTestStore(t)
-		ctx := context.Background()
-
-		s1 := extraction.NewStage1Filter(defaultExtractionConfig(), testLogger())
-		// Stage2/3 should never be called, but we need them for pipeline construction.
-		embedder := newPredictableEmbedder(3)
-		llm := &predictableLLM{rubricResponse: goodRubricJSON(), criticResponse: extractVerdictJSON()}
-		s2 := extraction.NewStage2Scorer(embedder, &mockQuerier{sim: 0.5}, llm, defaultExtractionConfig(), testLogger())
-		s3 := extraction.NewStage3Critic(llm, defaultExtractionConfig(), testLogger())
-
-		pipeline, _ := extraction.NewPipeline(extraction.PipelineConfig{
-			Stage1: s1, Stage2: s2, Stage3: s3, Writer: store, Log: testLogger(),
-		})
-
-		session := shortSession("sess-rej1", "test-lib") // too short → fails stage1
-		result, err := pipeline.Extract(ctx, session, []byte("trivial"))
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		if result.Status != model.StatusRejected {
-			t.Fatalf("status = %q, want rejected", result.Status)
-		}
-		if result.Stage1 == nil || result.Stage1.Passed {
-			t.Error("stage1 should have failed")
-		}
-		if result.Stage2 != nil {
-			t.Error("stage2 should not have been called")
-		}
-
-		// No skills should exist in DB.
-		results, err := store.Query(ctx, storage.SkillQuery{LibraryIDs: []string{"test-lib"}, Limit: 10})
-		if err != nil {
-			t.Fatal(err)
-		}
-		if len(results) != 0 {
-			t.Errorf("expected 0 skills in DB, got %d", len(results))
-		}
-	})
-
-	t.Run("rejected at stage3", func(t *testing.T) {
-		store := newTestStore(t)
-		ctx := context.Background()
-
-		embedder := newPredictableEmbedder(3)
-		llm := &predictableLLM{
-			rubricResponse: goodRubricJSON(),
-			criticResponse: rejectVerdictJSON(),
-		}
-
-		s1 := extraction.NewStage1Filter(defaultExtractionConfig(), testLogger())
-		s2 := extraction.NewStage2Scorer(embedder, &mockQuerier{sim: 0.5}, llm, defaultExtractionConfig(), testLogger())
-		s3 := extraction.NewStage3Critic(llm, defaultExtractionConfig(), testLogger())
-
-		pipeline, _ := extraction.NewPipeline(extraction.PipelineConfig{
-			Stage1: s1, Stage2: s2, Stage3: s3, Writer: store, Log: testLogger(),
-		})
-
-		session := goodSession("sess-rej3", "test-lib")
-		result, err := pipeline.Extract(ctx, session, []byte("This session fixed a complex database connection pooling issue."))
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		if result.Status != model.StatusRejected {
-			t.Fatalf("status = %q, want rejected", result.Status)
-		}
-		if result.Stage3 == nil || result.Stage3.Passed {
-			t.Error("stage3 should have rejected")
-		}
-		if result.Skill != nil {
-			t.Error("no skill should exist after stage3 rejection")
-		}
-
-		// DB should be empty.
-		results, _ := store.Query(ctx, storage.SkillQuery{LibraryIDs: []string{"test-lib"}, Limit: 10})
-		if len(results) != 0 {
-			t.Errorf("expected 0 skills, got %d", len(results))
-		}
-	})
-}
-
-// =============================================================================
-// Test 7: Error Recovery — Embedding Dies Mid-Pipeline
-// =============================================================================
-
-func TestEmbeddingFailureDegradation(t *testing.T) {
-	store := newTestStore(t)
-	ctx := context.Background()
-
-	// Embedder that always fails.
-	embedder := newPredictableEmbedder(3)
-	embedder.failAfter = -1 // always fail
-
-	llm := &predictableLLM{
-		rubricResponse: goodRubricJSON(),
-		criticResponse: extractVerdictJSON(),
-	}
-
-	s1 := extraction.NewStage1Filter(defaultExtractionConfig(), testLogger())
-	s2 := extraction.NewStage2Scorer(embedder, &mockQuerier{sim: 0.5}, llm, defaultExtractionConfig(), testLogger())
-	s3 := extraction.NewStage3Critic(llm, defaultExtractionConfig(), testLogger())
-
-	pipeline, _ := extraction.NewPipeline(extraction.PipelineConfig{
-		Stage1: s1, Stage2: s2, Stage3: s3, Writer: store, Log: testLogger(),
-	})
-
-	session := goodSession("sess-err-1", "test-lib")
-	result, err := pipeline.Extract(ctx, session, []byte("fix database issue"))
-
-	// Pipeline should not return a Go error — it wraps errors in result.
-	if err != nil {
-		t.Fatalf("pipeline returned error: %v", err)
-	}
-
-	// Should be StatusError since embedding failed in Stage2.
-	if result.Status != model.StatusError {
-		// Stage2 depends on embedding — if embed fails, Score returns error.
-		// Pipeline wraps this as StatusError.
-		t.Errorf("status = %q, want error (embedding failure should propagate as stage2 error)", result.Status)
-	}
-	if result.Status == model.StatusError && result.Error == "" {
-		t.Error("expected error message in result")
-	}
-
-	// No skills should be stored.
-	results, _ := store.Query(ctx, storage.SkillQuery{LibraryIDs: []string{"test-lib"}, Limit: 10})
-	if len(results) != 0 {
-		t.Errorf("expected 0 skills after failure, got %d", len(results))
-	}
-
-	// Test injection fallback: with embedding failure, injection uses pattern-only mode.
-	// Seed a skill for injection.
-	sk := &model.SkillRecord{
-		ID: "skill-fallback", Name: "fix-db", Version: 1, LibraryID: "test-lib",
-		Category: model.CategoryTactical,
-		Patterns: []string{"FIX/Backend/DatabaseConnection"},
-		Quality: model.QualityScores{
-			ProblemSpecificity: 4, SolutionCompleteness: 4, ContextPortability: 3,
-			ReasoningTransparency: 3, TechnicalAccuracy: 4, VerificationEvidence: 3,
-			InnovationLevel: 3, CompositeScore: 3.5, CriticConfidence: 0.8,
-		},
-		DecayScore:  1.0,
-		ExtractedBy: "test",
-		FilePath:    "skills/fix-db/SKILL.md",
-	}
-	if err := store.Put(ctx, sk, nil); err != nil {
-		t.Fatal(err)
-	}
-
-	// Injection with broken embedder (nil).
-	classifyLLM := &predictableLLM{
-		classifyResponse: classifyJSON("FIX", "Backend", []string{"FIX/Backend/DatabaseConnection"}),
-	}
-	inj := injection.New(nil, store, classifyLLM, nil, testLogger()) // nil embedder = degraded
-
-	resp, err := inj.Inject(ctx, model.InjectionRequest{
-		Prompt:     "fix db connection",
-		LibraryIDs: []string{"test-lib"},
-		MaxSkills:  3,
-	})
-	if err != nil {
-		t.Fatalf("injection should work in degraded mode: %v", err)
-	}
-	if len(resp.Skills) == 0 {
-		t.Error("pattern-only injection should still return matching skills")
-	}
-}
-
-// =============================================================================
-// Test 8: Stats Accuracy
-// Run known workload → verify metrics match expected values
-// =============================================================================
-
-func TestStatsAccuracy(t *testing.T) {
-	store := newTestStore(t)
-	ctx := context.Background()
-	embedder := newPredictableEmbedder(3)
-
-	llmExtract := &predictableLLM{
-		rubricResponse: goodRubricJSON(),
-		criticResponse: extractVerdictJSON(),
-	}
-
-	s1 := extraction.NewStage1Filter(defaultExtractionConfig(), testLogger())
-	s2 := extraction.NewStage2Scorer(embedder, &mockQuerier{sim: 0.5}, llmExtract, defaultExtractionConfig(), testLogger())
-	s3 := extraction.NewStage3Critic(llmExtract, defaultExtractionConfig(), testLogger())
-
-	indexer := storage.NewSQLiteIndexer(store)
-	committer := &indexingCommitter{indexer: indexer, embedder: embedder}
-	pipeline, _ := extraction.NewPipeline(extraction.PipelineConfig{
-		Stage1: s1, Stage2: s2, Stage3: s3, Committer: committer, Log: testLogger(),
-	})
-
-	// Process sessions: extract 1 good + reject 2 at stage1.
-	// Note: multiple extractions with the same rubric produce the same skill name,
-	// causing duplicate errors. This is realistic — each unique skill needs unique patterns.
-	sess0 := goodSession("sess-stats-good-0", "test-lib")
-	result0, _ := pipeline.Extract(ctx, sess0, []byte("fix database connection pooling"))
-	if result0.Status != model.StatusExtracted {
-		t.Fatalf("good session 0: status=%q error=%s", result0.Status, result0.Error)
-	}
-	insertSession(t, store.DB(), sess0, result0)
-
-	// Two more good sessions that will fail as duplicates (StatusError)
-	for i := 1; i < 3; i++ {
-		sess := goodSession(fmt.Sprintf("sess-stats-good-%d", i), "test-lib")
-		result, _ := pipeline.Extract(ctx, sess, []byte(fmt.Sprintf("fix variant %d", i)))
-		insertSession(t, store.DB(), sess, result)
-	}
-
-	for i := 0; i < 2; i++ {
-		sess := shortSession(fmt.Sprintf("sess-stats-bad-%d", i), "test-lib")
-		result, _ := pipeline.Extract(ctx, sess, []byte("tiny"))
-
-		insertSession(t, store.DB(), sess, result)
-	}
-
-	// Compute metrics.
-	collector := metrics.NewCollector(store.DB())
-	m, err := collector.Collect()
-	if err != nil {
-		t.Fatalf("collect metrics: %v", err)
-	}
-
-	if m.TotalSessions != 5 {
-		t.Errorf("total sessions = %d, want 5", m.TotalSessions)
-	}
-	// 1 extracted + 2 duplicates (error) + 2 rejected
-	if m.Extracted != 1 {
-		t.Errorf("extracted = %d, want 1", m.Extracted)
-	}
-	if m.Rejected != 2 {
-		t.Errorf("rejected = %d, want 2", m.Rejected)
-	}
-	if m.Errored != 2 {
-		t.Errorf("errored = %d, want 2", m.Errored)
-	}
-}
-
-// insertSession inserts a session record into the sessions table for metrics.
-func insertSession(t *testing.T, db *sql.DB, sess model.SessionRecord, result *model.ExtractionResult) {
-	t.Helper()
-	status := string(result.Status)
-	rejStage := 0
-	rejReason := ""
-	skillID := ""
-
-	if result.Status == model.StatusRejected {
-		switch {
-		case result.Stage1 != nil && !result.Stage1.Passed:
-			rejStage = 1
-			rejReason = result.Stage1.Reason
-		case result.Stage2 != nil && !result.Stage2.Passed:
-			rejStage = 2
-			rejReason = result.Stage2.Reason
-		case result.Stage3 != nil && !result.Stage3.Passed:
-			rejStage = 3
-			rejReason = result.Stage3.CriticReasoning
-		}
-	}
-	if result.Skill != nil {
-		skillID = result.Skill.ID
-	}
-
-	_, err := db.Exec(`INSERT INTO sessions (id, started_at, ended_at, duration_minutes, tool_call_count,
-		error_count, message_count, error_rate, has_successful_exec, tokens_used, agent_model, user_id,
-		library_id, extraction_status, rejected_at_stage, rejection_reason, extracted_skill_id)
-		VALUES (?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?)`,
-		sess.ID, sess.StartedAt, sess.EndedAt, sess.DurationMinutes, sess.ToolCallCount,
-		sess.ErrorCount, sess.MessageCount, sess.ErrorRate, sess.HasSuccessfulExec,
-		sess.TokensUsed, sess.AgentModel, sess.UserID,
-		sess.LibraryID, status, rejStage, rejReason,
-		sql.NullString{String: skillID, Valid: skillID != ""})
-	if err != nil {
-		t.Fatalf("insert session %s: %v", sess.ID, err)
-	}
-}
-
-// =============================================================================
-// Test 9: Decay Rescue — recently used skill gets rescued
-// =============================================================================
-
-func TestDecayRescue(t *testing.T) {
-	store := newTestStore(t)
-	ctx := context.Background()
-
-	now := time.Date(2026, 2, 15, 12, 0, 0, 0, time.UTC)
-
-	sk := &model.SkillRecord{
-		ID: "skill-rescue", Name: "rescue-me", Version: 1, LibraryID: "test-lib",
-		Category: model.CategoryTactical,
-		Patterns: []string{"FIX/Backend/DatabaseConnection"},
-		Quality: model.QualityScores{
-			ProblemSpecificity: 4, SolutionCompleteness: 4, ContextPortability: 3,
-			ReasoningTransparency: 3, TechnicalAccuracy: 4, VerificationEvidence: 3,
-			InnovationLevel: 3, CompositeScore: 3.5, CriticConfidence: 0.8,
-		},
-		DecayScore:         0.4, // already somewhat decayed
-		SuccessCorrelation: 0.8, // positive correlation
-		ExtractedBy:        "test",
-		FilePath:           "skills/rescue-me/SKILL.md",
-	}
-	if err := store.Put(ctx, sk, nil); err != nil {
-		t.Fatal(err)
-	}
-
-	// Set last_injected_at to 5 days ago (within rescue window).
-	lastInjected := now.AddDate(0, 0, -5)
-	store.DB().Exec("UPDATE skills SET last_injected_at = ?, success_correlation = 0.8 WHERE id = ?", lastInjected, "skill-rescue")
-
-	cfg := decay.DefaultConfig()
-	runner, err := decay.NewRunner(store, store, cfg, testLogger())
-	if err != nil {
-		t.Fatal(err)
-	}
-	runner.SetNow(func() time.Time { return now })
-
-	result, err := runner.RunCycle(ctx, "test-lib")
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if result.Rescued != 1 {
-		t.Errorf("rescued = %d, want 1", result.Rescued)
-	}
-
-	got, _ := store.Get(ctx, "skill-rescue")
-	// Decay: 0.4 * 0.5^(5/90) ≈ 0.385, + rescue boost 0.3 ≈ 0.685
-	if got.DecayScore < 0.6 {
-		t.Errorf("rescued decay = %.4f, want > 0.6", got.DecayScore)
-	}
-}
-
-// =============================================================================
-// Test 10: Multiple Extractions + Query Ordering
-// Verify that composite score ordering works with real DB data.
-// =============================================================================
-
-func TestMultipleSkillsQueryOrdering(t *testing.T) {
-	store := newTestStore(t)
-	ctx := context.Background()
-	embedder := newPredictableEmbedder(3)
-
-	// Create skills with varying quality and success correlation.
-	for i, tc := range []struct {
-		name        string
-		composite   float64
-		successCorr float64
-		patterns    []string
-	}{
-		{"low-quality", 2.0, -0.5, []string{"FIX/Backend/DatabaseConnection"}},
-		{"mid-quality", 3.5, 0.3, []string{"FIX/Backend/DatabaseConnection"}},
-		{"high-quality", 4.5, 0.8, []string{"FIX/Backend/DatabaseConnection"}},
-	} {
-		sk := &model.SkillRecord{
-			ID: fmt.Sprintf("skill-order-%d", i), Name: tc.name, Version: 1,
-			LibraryID: "test-lib", Category: model.CategoryTactical,
-			Patterns: tc.patterns,
-			Quality: model.QualityScores{
-				ProblemSpecificity: 4, SolutionCompleteness: 4, ContextPortability: 3,
-				ReasoningTransparency: 3, TechnicalAccuracy: 4, VerificationEvidence: 3,
-				InnovationLevel: 3, CompositeScore: tc.composite, CriticConfidence: 0.8,
-			},
-			Embedding:          embedder.deterministicVector(tc.name),
-			DecayScore:         1.0,
-			SuccessCorrelation: tc.successCorr,
-			ExtractedBy:        "test",
-			FilePath:           fmt.Sprintf("skills/%s/SKILL.md", tc.name),
-		}
-		if err := store.Put(ctx, sk, nil); err != nil {
-			t.Fatal(err)
-		}
-	}
-
-	// Query with pattern matching.
-	results, err := store.Query(ctx, storage.SkillQuery{
-		Patterns:   []string{"FIX/Backend/DatabaseConnection"},
-		LibraryIDs: []string{"test-lib"},
-		Limit:      10,
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if len(results) != 3 {
-		t.Fatalf("got %d results, want 3", len(results))
-	}
-
-	// Verify descending composite score order.
-	for i := 1; i < len(results); i++ {
-		if results[i].CompositeScore > results[i-1].CompositeScore {
-			t.Errorf("results not sorted: [%d].score=%f > [%d].score=%f",
-				i, results[i].CompositeScore, i-1, results[i-1].CompositeScore)
-		}
-	}
-
-	// The skill with highest success_correlation should have highest historical rate component.
-	// HistoricalRate = (success_correlation + 1) / 2
-	// high-quality: (0.8+1)/2 = 0.9
-	// low-quality: (-0.5+1)/2 = 0.25
-	best := results[0]
-	worst := results[len(results)-1]
-	if best.HistoricalRate <= worst.HistoricalRate {
-		t.Errorf("best historical rate %.2f should exceed worst %.2f", best.HistoricalRate, worst.HistoricalRate)
-	}
-}
-
-// =============================================================================
-// Test 11: Version Chain Integrity
-// Verify that skill versioning preserves parent chain in real DB.
-// =============================================================================
-
-func TestVersionChainIntegrity(t *testing.T) {
-	store := newTestStore(t)
-	ctx := context.Background()
-
-	v1 := &model.SkillRecord{
-		ID: "v1-id", Name: "fix-db", Version: 1, LibraryID: "test-lib",
-		Category: model.CategoryTactical,
-		Quality: model.QualityScores{
-			ProblemSpecificity: 3, SolutionCompleteness: 3, ContextPortability: 3,
-			ReasoningTransparency: 3, TechnicalAccuracy: 3, VerificationEvidence: 3,
-			InnovationLevel: 3, CompositeScore: 3.0, CriticConfidence: 0.7,
-		},
-		DecayScore: 1.0, ExtractedBy: "test", FilePath: "skills/fix-db/v1/SKILL.md",
-	}
-	if err := store.Put(ctx, v1, nil); err != nil {
-		t.Fatal(err)
-	}
-
-	v2 := &model.SkillRecord{
-		ID: "v2-id", Name: "fix-db", Version: 2, ParentID: "v1-id", LibraryID: "test-lib",
-		Category: model.CategoryTactical,
-		Quality: model.QualityScores{
-			ProblemSpecificity: 4, SolutionCompleteness: 4, ContextPortability: 3,
-			ReasoningTransparency: 3, TechnicalAccuracy: 4, VerificationEvidence: 3,
-			InnovationLevel: 3, CompositeScore: 3.5, CriticConfidence: 0.8,
-		},
-		DecayScore: 1.0, ExtractedBy: "test", FilePath: "skills/fix-db/v2/SKILL.md",
-	}
-	if err := store.Put(ctx, v2, nil); err != nil {
-		t.Fatal(err)
-	}
-
-	// GetLatestByName should return v2.
-	latest, err := store.GetLatestByName(ctx, "fix-db", "test-lib")
-	if err != nil {
-		t.Fatal(err)
-	}
-	if latest.Version != 2 {
-		t.Errorf("latest version = %d, want 2", latest.Version)
-	}
-	if latest.ParentID != "v1-id" {
-		t.Errorf("parent = %q, want v1-id", latest.ParentID)
-	}
-
-	// Verify v1 still exists.
-	old, err := store.Get(ctx, "v1-id")
-	if err != nil {
-		t.Fatal(err)
-	}
-	if old.Version != 1 {
-		t.Errorf("v1 version = %d", old.Version)
-	}
-}
-
-// =============================================================================
-// Test 12: Dead Skill Filtering in Injection
-// Deprecated skills (decay=0) should not appear in injection results.
-// =============================================================================
-
-func TestDeadSkillFiltering(t *testing.T) {
-	store := newTestStore(t)
-	ctx := context.Background()
-	embedder := newPredictableEmbedder(3)
-
-	// One alive skill, one dead.
-	alive := &model.SkillRecord{
-		ID: "alive", Name: "alive-skill", Version: 1, LibraryID: "test-lib",
-		Category: model.CategoryTactical,
-		Patterns: []string{"FIX/Backend/DatabaseConnection"},
-		Quality: model.QualityScores{
-			ProblemSpecificity: 4, SolutionCompleteness: 4, ContextPortability: 3,
-			ReasoningTransparency: 3, TechnicalAccuracy: 4, VerificationEvidence: 3,
-			InnovationLevel: 3, CompositeScore: 3.5, CriticConfidence: 0.8,
-		},
-		Embedding:   embedder.deterministicVector("alive skill"),
-		DecayScore:  0.8,
-		ExtractedBy: "test",
-		FilePath:    "skills/alive/SKILL.md",
-	}
-	dead := &model.SkillRecord{
-		ID: "dead", Name: "dead-skill", Version: 1, LibraryID: "test-lib",
-		Category: model.CategoryTactical,
-		Patterns: []string{"FIX/Backend/DatabaseConnection"},
-		Quality: model.QualityScores{
-			ProblemSpecificity: 4, SolutionCompleteness: 4, ContextPortability: 3,
-			ReasoningTransparency: 3, TechnicalAccuracy: 4, VerificationEvidence: 3,
-			InnovationLevel: 3, CompositeScore: 3.5, CriticConfidence: 0.8,
-		},
-		Embedding:   embedder.deterministicVector("dead skill"),
-		DecayScore:  0.0,
-		ExtractedBy: "test",
-		FilePath:    "skills/dead/SKILL.md",
-	}
-
-	store.Put(ctx, alive, nil)
-	store.Put(ctx, dead, nil)
-
-	llm := &predictableLLM{
-		classifyResponse: classifyJSON("FIX", "Backend", []string{"FIX/Backend/DatabaseConnection"}),
-	}
-	inj := injection.New(embedder, store, llm, store, testLogger())
-
-	resp, err := inj.Inject(ctx, model.InjectionRequest{
-		Prompt:     "fix db",
-		LibraryIDs: []string{"test-lib"},
-		MaxSkills:  10,
-		SessionID:  "sess-dead-test",
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	for _, s := range resp.Skills {
-		if s.SkillID == "dead" {
-			t.Error("dead skill (decay=0) should not appear in injection results")
-		}
-	}
-	if len(resp.Skills) == 0 {
-		t.Error("alive skill should still appear")
-	}
-}
-
-// =============================================================================
-// Test 13: Transaction Boundary — Skill Put Atomicity
-// If pattern insert fails, skill row should also roll back.
-// =============================================================================
-
-func TestSkillPutAtomicity(t *testing.T) {
-	store := newTestStore(t)
-	ctx := context.Background()
-
-	// First insert succeeds.
-	sk1 := &model.SkillRecord{
-		ID: "atom-1", Name: "atomic-skill", Version: 1, LibraryID: "test-lib",
-		Category: model.CategoryTactical,
-		Patterns: []string{"FIX/Backend/DatabaseConnection"},
-		Quality: model.QualityScores{
-			ProblemSpecificity: 4, SolutionCompleteness: 4, ContextPortability: 3,
-			ReasoningTransparency: 3, TechnicalAccuracy: 4, VerificationEvidence: 3,
-			InnovationLevel: 3, CompositeScore: 3.5, CriticConfidence: 0.8,
-		},
-		DecayScore: 1.0, ExtractedBy: "test", FilePath: "skills/atomic/SKILL.md",
-	}
-	if err := store.Put(ctx, sk1, nil); err != nil {
-		t.Fatal(err)
-	}
-
-	// Second insert with same name+version+library should fail (unique constraint).
-	sk2 := &model.SkillRecord{
-		ID: "atom-2", Name: "atomic-skill", Version: 1, LibraryID: "test-lib",
-		Category: model.CategoryTactical,
-		Patterns: []string{"BUILD/Backend/APIDesign"},
-		Quality: model.QualityScores{
-			ProblemSpecificity: 3, SolutionCompleteness: 3, ContextPortability: 3,
-			ReasoningTransparency: 3, TechnicalAccuracy: 3, VerificationEvidence: 3,
-			InnovationLevel: 3, CompositeScore: 3.0, CriticConfidence: 0.7,
-		},
-		DecayScore: 1.0, ExtractedBy: "test", FilePath: "skills/atomic/SKILL.md",
-	}
-	err := store.Put(ctx, sk2, nil)
-	if err == nil {
-		t.Fatal("expected duplicate error")
-	}
-
-	// Verify atom-2's patterns were NOT inserted (transaction rolled back).
-	var patternCount int
-	store.DB().QueryRow("SELECT COUNT(*) FROM skill_patterns WHERE skill_id = ?", "atom-2").Scan(&patternCount)
-	if patternCount != 0 {
-		t.Errorf("atom-2 patterns = %d, want 0 (should have rolled back)", patternCount)
-	}
-
-	// Original skill should still be intact.
-	got, err := store.Get(ctx, "atom-1")
-	if err != nil {
-		t.Fatal(err)
-	}
-	if got.Name != "atomic-skill" {
-		t.Errorf("original skill corrupted: name = %q", got.Name)
-	}
-}
-
-// =============================================================================
-// Test 14: Concurrent Extraction Safety
-// Multiple extractions writing to same DB should not corrupt data.
-// =============================================================================
-
-func TestConcurrentExtractions(t *testing.T) {
-	// Use file-based temp DB for concurrent access (in-memory has issues with concurrent goroutines).
-	tmpDir := t.TempDir()
-	store, err := storage.NewSQLiteStore(tmpDir+"/concurrent.db", "test-model")
-	if err != nil {
-		t.Fatal(err)
-	}
-	t.Cleanup(func() { store.Close() })
-	ctx := context.Background()
-
-	const n = 10
-	errs := make(chan error, n)
-
-	for i := 0; i < n; i++ {
-		go func(idx int) {
-			embedder := newPredictableEmbedder(3)
-			llm := &predictableLLM{
-				rubricResponse: goodRubricJSON(),
-				criticResponse: extractVerdictJSON(),
-			}
-
-			s1 := extraction.NewStage1Filter(defaultExtractionConfig(), testLogger())
-			s2 := extraction.NewStage2Scorer(embedder, &mockQuerier{sim: 0.5}, llm, defaultExtractionConfig(), testLogger())
-			s3 := extraction.NewStage3Critic(llm, defaultExtractionConfig(), testLogger())
-
-			indexer := storage.NewSQLiteIndexer(store)
-			committer := &indexingCommitter{indexer: indexer, embedder: embedder}
-			pipeline, _ := extraction.NewPipeline(extraction.PipelineConfig{
-				Stage1: s1, Stage2: s2, Stage3: s3, Committer: committer, Log: testLogger(),
-			})
-
-			sess := goodSession(fmt.Sprintf("sess-conc-%d", idx), "test-lib")
-			content := []byte(fmt.Sprintf("unique problem %d requiring specialized solution", idx))
-			result, err := pipeline.Extract(ctx, sess, content)
-			if err != nil {
-				errs <- err
-				return
-			}
-			if result.Status == model.StatusError {
-				errs <- fmt.Errorf("session %d: %s", idx, result.Error)
-				return
-			}
-			errs <- nil
-		}(i)
-	}
-
-	var failures int
-	for i := 0; i < n; i++ {
-		if err := <-errs; err != nil {
-			t.Logf("concurrent extraction %d: %v", i, err)
-			failures++
-		}
-	}
-
-	// Many will fail due to unique name collisions (same pattern → same name).
-	// The key test: no panics, no DB corruption.
-	results, err := store.Query(ctx, storage.SkillQuery{LibraryIDs: []string{"test-lib"}, Limit: 100})
-	if err != nil {
-		t.Fatalf("query after concurrent: %v", err)
-	}
-	// At least one should succeed (first one to commit).
-	if len(results) == 0 {
-		t.Error("expected at least one skill from concurrent extraction")
-	}
-
-	// Verify DB integrity after concurrent writes.
-	var integrity string
-	store.DB().QueryRow("PRAGMA integrity_check").Scan(&integrity)
-	if integrity != "ok" {
-		t.Errorf("DB integrity: %s", integrity)
-	}
-}
-
-// =============================================================================
-// Test 15: Schema Constraint Validation
-// Verify CHECK constraints work with real SQLite.
-// =============================================================================
-
-func TestSchemaConstraints(t *testing.T) {
-	store := newTestStore(t)
-	ctx := context.Background()
-
-	t.Run("invalid category rejected", func(t *testing.T) {
-		sk := &model.SkillRecord{
-			ID: "bad-cat", Name: "bad-category", Version: 1, LibraryID: "test-lib",
-			Category: model.SkillCategory("invalid"),
-			Quality: model.QualityScores{
-				ProblemSpecificity: 4, SolutionCompleteness: 4, ContextPortability: 3,
-				ReasoningTransparency: 3, TechnicalAccuracy: 4, VerificationEvidence: 3,
-				InnovationLevel: 3, CompositeScore: 3.5, CriticConfidence: 0.8,
-			},
-			DecayScore: 1.0, ExtractedBy: "test", FilePath: "skills/bad/SKILL.md",
-		}
-		err := store.Put(ctx, sk, nil)
-		if err == nil {
-			t.Error("expected error for invalid category")
-		}
-	})
-
-	t.Run("quality score out of range rejected", func(t *testing.T) {
-		sk := &model.SkillRecord{
-			ID: "bad-score", Name: "bad-score", Version: 1, LibraryID: "test-lib",
-			Category: model.CategoryTactical,
-			Quality: model.QualityScores{
-				ProblemSpecificity:   10, // out of range!
-				SolutionCompleteness: 4, ContextPortability: 3,
-				ReasoningTransparency: 3, TechnicalAccuracy: 4, VerificationEvidence: 3,
-				InnovationLevel: 3, CompositeScore: 3.5, CriticConfidence: 0.8,
-			},
-			DecayScore: 1.0, ExtractedBy: "test", FilePath: "skills/bad/SKILL.md",
-		}
-		err := store.Put(ctx, sk, nil)
-		if err == nil {
-			t.Error("expected error for quality score out of range")
-		}
-	})
-
-	t.Run("confidence out of range rejected", func(t *testing.T) {
-		sk := &model.SkillRecord{
-			ID: "bad-conf", Name: "bad-conf", Version: 1, LibraryID: "test-lib",
-			Category: model.CategoryTactical,
-			Quality: model.QualityScores{
-				ProblemSpecificity: 4, SolutionCompleteness: 4, ContextPortability: 3,
-				ReasoningTransparency: 3, TechnicalAccuracy: 4, VerificationEvidence: 3,
-				InnovationLevel: 3, CompositeScore: 3.5, CriticConfidence: 1.5, // out of range!
-			},
-			DecayScore: 1.0, ExtractedBy: "test", FilePath: "skills/bad/SKILL.md",
-		}
-		err := store.Put(ctx, sk, nil)
-		if err == nil {
-			t.Error("expected error for confidence out of range")
-		}
-	})
-}
-
-// =============================================================================
-// Test 16: Session Persistence Through Full Pipeline (Jazz N4)
-// =============================================================================
-
-func TestSessionPersistence(t *testing.T) {
-	store := newTestStore(t)
-	ctx := context.Background()
-	embedder := newPredictableEmbedder(3)
-
-	llm := &predictableLLM{
-		rubricResponse: goodRubricJSON(),
-		criticResponse: extractVerdictJSON(),
-	}
-
-	s1 := extraction.NewStage1Filter(defaultExtractionConfig(), testLogger())
-	s2 := extraction.NewStage2Scorer(embedder, &mockQuerier{sim: 0.50}, llm, defaultExtractionConfig(), testLogger())
-	s3 := extraction.NewStage3Critic(llm, defaultExtractionConfig(), testLogger())
-
-	indexer := storage.NewSQLiteIndexer(store)
-	committer := &indexingCommitter{indexer: indexer, embedder: embedder}
-	pipeline, err := extraction.NewPipeline(extraction.PipelineConfig{
-		Stage1:     s1,
-		Stage2:     s2,
-		Stage3:     s3,
-		Committer:  committer,
-		Sessions:   store, // <-- wired!
-		Embedder:   embedder,
-		Log:        testLogger(),
-		SampleRate: 0,
-		Extractor:  "session-persist-test",
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Test 1: Successful extraction — session should be persisted with status=extracted.
-	sess := goodSession("sess-persist-1", "test-lib")
-	result, err := pipeline.Extract(ctx, sess, []byte("fix database connection pooling issue with retry logic"))
-	if err != nil {
-		t.Fatal(err)
-	}
-	if result.Status != model.StatusExtracted {
-		t.Fatalf("status = %q, want extracted", result.Status)
-	}
-
-	// Verify session row exists in DB.
-	var status string
-	var skillID sql.NullString
-	err = store.DB().QueryRow("SELECT extraction_status, extracted_skill_id FROM sessions WHERE id = ?", "sess-persist-1").Scan(&status, &skillID)
-	if err != nil {
-		t.Fatalf("query session: %v", err)
-	}
-	if status != "extracted" {
-		t.Errorf("session status = %q, want extracted", status)
-	}
-	if !skillID.Valid || skillID.String == "" {
-		t.Error("extracted_skill_id should be set")
-	}
-
-	// Test 2: Rejected session — session should be persisted with status=rejected.
-	rejSess := shortSession("sess-persist-2", "test-lib")
-	rejResult, err := pipeline.Extract(ctx, rejSess, []byte("tiny"))
-	if err != nil {
-		t.Fatal(err)
-	}
-	if rejResult.Status != model.StatusRejected {
-		t.Fatalf("status = %q, want rejected", rejResult.Status)
-	}
-
-	var rejStatus string
-	var rejStage int
-	var rejReason string
-	err = store.DB().QueryRow("SELECT extraction_status, rejected_at_stage, rejection_reason FROM sessions WHERE id = ?", "sess-persist-2").Scan(&rejStatus, &rejStage, &rejReason)
-	if err != nil {
-		t.Fatalf("query rejected session: %v", err)
-	}
-	if rejStatus != "rejected" {
-		t.Errorf("rejected session status = %q, want rejected", rejStatus)
-	}
-	if rejStage != 1 {
-		t.Errorf("rejected_at_stage = %d, want 1", rejStage)
-	}
-	if rejReason == "" {
-		t.Error("rejection_reason should not be empty")
-	}
-}
-
-// =============================================================================
-// Test 17: Embedding Cosine Similarity E2E
-// Extract skill with embedding → inject → verify cosine sim computed
-// =============================================================================
-
-func TestEmbeddingCosineSimilarityE2E(t *testing.T) {
-	store := newTestStore(t)
-	ctx := context.Background()
-	embedder := newPredictableEmbedder(3)
-
-	// Seed a skill WITH embedding.
-	vec := embedder.deterministicVector("fix database connection")
-	sk := &model.SkillRecord{
-		ID: "skill-cos-1", Name: "fix-db-conn", Version: 1, LibraryID: "test-lib",
-		Category: model.CategoryTactical,
-		Patterns: []string{"FIX/Backend/DatabaseConnection"},
-		Quality: model.QualityScores{
-			ProblemSpecificity: 4, SolutionCompleteness: 4, ContextPortability: 3,
-			ReasoningTransparency: 3, TechnicalAccuracy: 4, VerificationEvidence: 3,
-			InnovationLevel: 3, CompositeScore: 3.5, CriticConfidence: 0.8,
-		},
-		Embedding:   vec,
-		DecayScore:  1.0,
-		ExtractedBy: "test",
-		FilePath:    "skills/fix-db/SKILL.md",
-	}
-	if err := store.Put(ctx, sk, nil); err != nil {
-		t.Fatal(err)
-	}
-
-	// Inject with a SIMILAR prompt (should produce similar embedding).
-	llm := &predictableLLM{
-		classifyResponse: classifyJSON("FIX", "Backend", []string{"FIX/Backend/DatabaseConnection"}),
-	}
-	inj := injection.New(embedder, store, llm, store, testLogger())
-
-	resp, err := inj.Inject(ctx, model.InjectionRequest{
-		Prompt:     "fix database connection", // same text = same embedding = cosine 1.0
-		LibraryIDs: []string{"test-lib"},
-		MaxSkills:  5,
-		SessionID:  "sess-cos-1",
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-	if len(resp.Skills) == 0 {
-		t.Fatal("expected injected skills")
-	}
-
-	// Cosine similarity should be ~1.0 for identical text.
-	if resp.Skills[0].CosineSim < 0.99 {
-		t.Errorf("cosine sim = %f, want ~1.0 for identical embedding", resp.Skills[0].CosineSim)
-	}
-
-	// Now inject with a DIFFERENT prompt.
-	resp2, err := inj.Inject(ctx, model.InjectionRequest{
-		Prompt:     "build REST API design patterns",
-		LibraryIDs: []string{"test-lib"},
-		MaxSkills:  5,
-		SessionID:  "sess-cos-2",
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-	if len(resp2.Skills) == 0 {
-		t.Fatal("expected skills even for different prompt (pattern match)")
-	}
-
-	// Cosine similarity should be lower for different text.
-	if resp2.Skills[0].CosineSim >= 0.99 {
-		t.Errorf("cosine sim = %f, should be < 0.99 for different embedding", resp2.Skills[0].CosineSim)
-	}
-}
-
-// =============================================================================
-// Test 18: A/B Event Deduplication — verify exact event count
-// =============================================================================
-
-func TestABEventDeduplication(t *testing.T) {
-	store := newTestStore(t)
-	ctx := context.Background()
-	embedder := newPredictableEmbedder(3)
-
-	// Seed a skill.
-	sk := &model.SkillRecord{
-		ID: "skill-dedup-1", Name: "fix-deploy", Version: 1, LibraryID: "test-lib",
-		Category: model.CategoryTactical,
-		Patterns: []string{"FIX/DevOps/DeploymentFailure"},
-		Quality: model.QualityScores{
-			ProblemSpecificity: 4, SolutionCompleteness: 4, ContextPortability: 3,
-			ReasoningTransparency: 3, TechnicalAccuracy: 4, VerificationEvidence: 3,
-			InnovationLevel: 3, CompositeScore: 3.5, CriticConfidence: 0.8,
-		},
-		Embedding:   embedder.deterministicVector("fix deployment failure"),
-		DecayScore:  1.0,
-		ExtractedBy: "test",
-		FilePath:    "skills/fix-deploy/SKILL.md",
-	}
-	if err := store.Put(ctx, sk, nil); err != nil {
-		t.Fatal(err)
-	}
-
-	llm := &predictableLLM{
-		classifyResponse: classifyJSON("FIX", "DevOps", []string{"FIX/DevOps/DeploymentFailure"}),
-	}
-
-	// Inner injector with nil eventWriter — AB handles events.
-	innerInj := injection.New(embedder, store, llm, nil, testLogger())
-	abInj := injection.NewABInjector(innerInj, store, injection.ABConfig{
-		Enabled: true, ControlRatio: 0.5, MinSampleSize: 10,
-	}, testLogger())
-
-	abInj.SetRandFunc(func() float64 { return 0.9 }) // treatment
-
-	_, err := abInj.Inject(ctx, model.InjectionRequest{
-		Prompt:     "fix deployment failure",
-		LibraryIDs: []string{"test-lib"},
-		MaxSkills:  5,
-		SessionID:  "sess-dedup-1",
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Verify EXACTLY 1 event per skill (no duplicates).
-	var totalEvents int
-	store.DB().QueryRow("SELECT COUNT(*) FROM injection_events WHERE session_id = ?", "sess-dedup-1").Scan(&totalEvents)
-	if totalEvents != 1 {
-		t.Errorf("total events = %d, want exactly 1 (no duplicates)", totalEvents)
-	}
-
-	// All events should have ab_group set.
-	var ungrouped int
-	store.DB().QueryRow("SELECT COUNT(*) FROM injection_events WHERE session_id = ? AND ab_group = ''", "sess-dedup-1").Scan(&ungrouped)
-	if ungrouped != 0 {
-		t.Errorf("ungrouped events = %d, want 0 (all should have ab_group)", ungrouped)
-	}
-}
-
-// =============================================================================
-// Test 19: Human Review Sampling E2E
-// =============================================================================
-
-func TestHumanReviewSampling(t *testing.T) {
-	store := newTestStore(t)
-	ctx := context.Background()
-	embedder := newPredictableEmbedder(3)
-
-	llm := &predictableLLM{
-		rubricResponse: goodRubricJSON(),
-		criticResponse: extractVerdictJSON(),
-	}
-
-	s1 := extraction.NewStage1Filter(defaultExtractionConfig(), testLogger())
-	s2 := extraction.NewStage2Scorer(embedder, &mockQuerier{sim: 0.50}, llm, defaultExtractionConfig(), testLogger())
-	s3 := extraction.NewStage3Critic(llm, defaultExtractionConfig(), testLogger())
-
-	pipeline, err := extraction.NewPipeline(extraction.PipelineConfig{
-		Stage1:     s1,
-		Stage2:     s2,
-		Stage3:     s3,
-		Writer:     store,
-		Sessions:   store,
-		Reviewer:   store, // real store, not mock
-		Embedder:   embedder,
-		Log:        testLogger(),
-		SampleRate: 1.0, // always sample
-		Extractor:  "sampling-test",
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Run extraction — should produce a sample.
-	sess := goodSession("sess-sample-1", "test-lib")
-	result, err := pipeline.Extract(ctx, sess, []byte("fix database connection pooling"))
-	if err != nil {
-		t.Fatal(err)
-	}
-	if result.Status != model.StatusExtracted {
-		t.Fatalf("status = %q, want extracted", result.Status)
-	}
-
-	// Verify human_review_samples row exists.
-	var sampleCount int
-	store.DB().QueryRow("SELECT COUNT(*) FROM human_review_samples WHERE session_id = ?", "sess-sample-1").Scan(&sampleCount)
-	if sampleCount != 1 {
-		t.Errorf("sample count = %d, want 1", sampleCount)
-	}
-
-	// Verify the sample contains valid JSON.
-	var resultJSON string
-	store.DB().QueryRow("SELECT extraction_result FROM human_review_samples WHERE session_id = ?", "sess-sample-1").Scan(&resultJSON)
-	if resultJSON == "" {
-		t.Error("extraction_result should not be empty")
-	}
-}
-
-// =============================================================================
-// Test 20: Stats Through Full Pipeline Chain
-// =============================================================================
-
-func TestStatsThroughPipeline(t *testing.T) {
-	store := newTestStore(t)
-	ctx := context.Background()
-	embedder := newPredictableEmbedder(3)
-
-	llm := &predictableLLM{
-		rubricResponse: goodRubricJSON(),
-		criticResponse: extractVerdictJSON(),
-	}
-
-	s1 := extraction.NewStage1Filter(defaultExtractionConfig(), testLogger())
-	s2 := extraction.NewStage2Scorer(embedder, &mockQuerier{sim: 0.50}, llm, defaultExtractionConfig(), testLogger())
-	s3 := extraction.NewStage3Critic(llm, defaultExtractionConfig(), testLogger())
-
-	indexer := storage.NewSQLiteIndexer(store)
-	committer := &indexingCommitter{indexer: indexer, embedder: embedder}
-	pipeline, err := extraction.NewPipeline(extraction.PipelineConfig{
-		Stage1: s1, Stage2: s2, Stage3: s3,
-		Committer: committer,
-		Sessions:  store,
-		Embedder:  embedder,
-		Log:       testLogger(),
-		Extractor: "stats-pipeline-test",
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// 1 good extraction.
-	sess1 := goodSession("sess-sp-1", "test-lib")
-	r1, _ := pipeline.Extract(ctx, sess1, []byte("fix database connection pooling"))
-	if r1.Status != model.StatusExtracted {
-		t.Fatalf("expected extracted, got %q: %s", r1.Status, r1.Error)
-	}
-
-	// 1 rejection at stage1.
-	sess2 := shortSession("sess-sp-2", "test-lib")
-	r2, _ := pipeline.Extract(ctx, sess2, []byte("tiny"))
-	if r2.Status != model.StatusRejected {
-		t.Fatalf("expected rejected, got %q", r2.Status)
-	}
-
-	// Collect metrics — should reflect pipeline-persisted sessions.
-	collector := metrics.NewCollector(store.DB())
-	m, err := collector.Collect()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if m.TotalSessions != 2 {
-		t.Errorf("total sessions = %d, want 2", m.TotalSessions)
-	}
-	if m.Extracted != 1 {
-		t.Errorf("extracted = %d, want 1", m.Extracted)
-	}
-	if m.Rejected != 1 {
-		t.Errorf("rejected = %d, want 1", m.Rejected)
-	}
-}
-
-// =============================================================================
-// Test 21: Concurrent Extraction + Injection (race conditions)
-// =============================================================================
-
-func TestConcurrentExtractionAndInjection(t *testing.T) {
-	tmpDir := t.TempDir()
-	store, err := storage.NewSQLiteStore(tmpDir+"/concurrent-ei.db", "test-model")
-	if err != nil {
-		t.Fatal(err)
-	}
-	t.Cleanup(func() { store.Close() })
-	ctx := context.Background()
-	embedder := newPredictableEmbedder(3)
-
-	// Pre-seed a skill for injection to find.
-	sk := &model.SkillRecord{
-		ID: "pre-seed-1", Name: "pre-seeded", Version: 1, LibraryID: "test-lib",
-		Category: model.CategoryTactical,
-		Patterns: []string{"FIX/Backend/DatabaseConnection"},
-		Quality: model.QualityScores{
-			ProblemSpecificity: 4, SolutionCompleteness: 4, ContextPortability: 3,
-			ReasoningTransparency: 3, TechnicalAccuracy: 4, VerificationEvidence: 3,
-			InnovationLevel: 3, CompositeScore: 3.5, CriticConfidence: 0.8,
-		},
-		Embedding:   embedder.deterministicVector("pre-seeded skill"),
-		DecayScore:  1.0,
-		ExtractedBy: "test",
-		FilePath:    "skills/pre-seeded/SKILL.md",
-	}
-	if err := store.Put(ctx, sk, nil); err != nil {
-		t.Fatal(err)
-	}
-
-	const n = 5
-	errs := make(chan error, n*2)
-
-	// Extraction goroutines.
-	for i := 0; i < n; i++ {
-		go func(idx int) {
-			emb := newPredictableEmbedder(3)
-			llm := &predictableLLM{rubricResponse: goodRubricJSON(), criticResponse: extractVerdictJSON()}
-			s1 := extraction.NewStage1Filter(defaultExtractionConfig(), testLogger())
-			s2 := extraction.NewStage2Scorer(emb, &mockQuerier{sim: 0.5}, llm, defaultExtractionConfig(), testLogger())
-			s3 := extraction.NewStage3Critic(llm, defaultExtractionConfig(), testLogger())
-			p, _ := extraction.NewPipeline(extraction.PipelineConfig{
-				Stage1: s1, Stage2: s2, Stage3: s3, Writer: store, Log: testLogger(),
-			})
-			sess := goodSession(fmt.Sprintf("sess-cei-%d", idx), "test-lib")
-			_, err := p.Extract(ctx, sess, []byte(fmt.Sprintf("unique extraction problem %d", idx)))
-			errs <- err
-		}(i)
-	}
-
-	// Injection goroutines.
-	for i := 0; i < n; i++ {
-		go func(idx int) {
-			emb := newPredictableEmbedder(3)
-			llm := &predictableLLM{classifyResponse: classifyJSON("FIX", "Backend", []string{"FIX/Backend/DatabaseConnection"})}
-			inj := injection.New(emb, store, llm, store, testLogger())
-			_, err := inj.Inject(ctx, model.InjectionRequest{
-				Prompt:     "fix database issue",
-				LibraryIDs: []string{"test-lib"},
-				MaxSkills:  3,
-				SessionID:  fmt.Sprintf("sess-cinj-%d", idx),
-			})
-			errs <- err
-		}(i)
-	}
-
-	var failures int
-	for i := 0; i < n*2; i++ {
-		if err := <-errs; err != nil {
-			t.Logf("concurrent op %d: %v", i, err)
-			failures++
-		}
-	}
-
-	// Key assertion: no panics, no corruption.
-	var integrity string
-	store.DB().QueryRow("PRAGMA integrity_check").Scan(&integrity)
-	if integrity != "ok" {
-		t.Fatalf("DB integrity: %s", integrity)
-	}
-}
-
-// math is used in helpers_test.go (assertApprox)
